fix(security): close the key-shaped redaction gap and make the slot digest injective
maskSecretsDeep walked for secret-shaped STRINGS, so a credential arriving as a
real JSON value under a secret-named key kept no keyword/separator context and
survived into the receipt. The key name is now sufficient on its own and masks
whatever hangs beneath it - object, array or number. Whole-word matching (camel
/snake/kebab split, depluralized, adjacent pairs joined) keeps tokenizer,
passwordless and unauthorized out of the secret set.

The idempotency slot preimage joined its three fields with a single delimiter,
which is not injective: a delimiter inside a field is indistinguishable from a
field boundary, so two distinct scopes collapsed onto one slot - and the two
mirrors did not even use the same delimiter. Both now length-prefix each field
(<utf8ByteLength> ':' <utf8Bytes>) and digest byte-identical preimages. Three
vectors computed with an external sha256sum are pinned on BOTH surfaces, so a
future divergence reds exactly one of the two suites. The raw key is still
hashed and never stored.

The constant-time token compare folded the length delta through a uint8 cast,
so any length pair whose XOR is a nonzero multiple of 256 (1 vs 257) truncated
to zero and lost the mismatch. The accumulator is uint32 now; still branch-free,
still no early exit.

The resources parity fixture ran two full transport captures on one 30s hook
and blew it under load while passing in isolation. The minimal-profile capture
is memoized and paid by its single consumer; the budget is sized for a contended
runner. No assertion changed.

Eviction stays COMPLETED-only and in-flight entries stay unbounded and
non-expiring on purpose: both a cap and a lease would have to release a slot
whose dispatch may still be running, which is the duplicate mutation the ledger
exists to prevent.

The native C++ changes are NOT compiled in this lane (no editor, no plugin
build, no RunUAT); they are static edits plus source-contract tests.

+++ plugins/McpAutomationBridge/Source/McpAutomationBridge/Private/Foundation/McpSecureTokenCompare.h
@@ -17,7 +17,10 @@
  *
  * Lengths are folded into the accumulator (LenA ^ LenB) rather than used to
  * short-circuit, so an empty or shorter candidate still compares unequal
- * without branching on the length difference.
+ * without branching on the length difference. The accumulator is uint32-wide
+ * because the length XOR does not fit in a byte: narrowing it truncates, and
+ * any length pair whose XOR is a nonzero multiple of 256 (1 vs 257, 16 vs 272)
+ * would fold to zero and lose the mismatch entirely.
  */
 inline bool McpConstantTimeTokenEquals(const FString& A, const FString& B)
 {
@@ -28,8 +31,9 @@ inline bool McpConstantTimeTokenEquals(const FString& A, const FString& B)
 	const int32 MaxLen = FMath::Max(LenA, LenB);
 
 	// Start the accumulator from the length difference so length mismatches
-	// are captured without an early-out branch.
-	uint8 Diff = static_cast<uint8>(LenA ^ LenB);
+	// are captured without an early-out branch. Full width, not uint8: the XOR
+	// of two lengths can carry its only set bits above bit 7.
+	uint32 Diff = static_cast<uint32>(LenA ^ LenB);
 
 	const uint8* Pa = reinterpret_cast<const uint8*>(Utf8A.Get());
 	const uint8* Pb = reinterpret_cast<const uint8*>(Utf8B.Get());
